탐색 메인 페이지 및 Search 컴포넌트 수정
