+++ src/components/Search/Search.tsx
@@ -2,15 +2,24 @@ import React from "react";
 import styled from "styled-components";
 import SearchIcon from "../../assets/icons/searchIcon.svg";
 
-export default function Search({ placeholder, value, onChange, ...props }) {
+// export default function Search({ placeholder, value, onChange, ...props }) {
+export default function Search({ placeholder, value, onKeyDown, ...props }) {
+  const [isInputFocused, setIsInputFocused] = React.useState<boolean>(false);
+  const [inputValue, setInputValue] = React.useState<string>("");
+
   return (
     <SearchWrapper {...props}>
       <img src={SearchIcon} alt="Search Icon" />
       <input
         type="text"
-        placeholder={placeholder}
-        value={value}
-        onChange={onChange}
+        // placeholder={placeholder}
+        placeholder={!isInputFocused ? placeholder : ''}
+        value={inputValue}
+        // onChange={onChange}
+        onFocus={() => setIsInputFocused(true)}
+        onBlur={() => setIsInputFocused(false)}
+        onChange={(e) => setInputValue(e.target.value)}
+        onKeyDown={onKeyDown}
       />
     </SearchWrapper>
   );
@@ -20,7 +29,7 @@ const SearchWrapper = styled.div`
   display: flex;
   align-items: center;
   border-radius: 10px;
-  padding: 5px 12px;
+  padding: 8px 17px;
   width: 100%;
   background-color: #eaeaea;
   gap: 8px;
@@ -35,13 +44,25 @@ const SearchWrapper = styled.div`
     flex: 1;
     background-color: inherit;
     border: none;
-    font-size: 16px;
+    color: #C1C1C1;
+    font-feature-settings: 'liga' off, 'clig' off;
+    font-family: Pretendard;
+    font-size: 17px;
+    font-style: normal;
     font-weight: 500;
-    color: #333;
+    line-height: 140%; /* 23.8px */
+    letter-spacing: -0.425px;
   }
 
   > input::placeholder {
-    color: #c0c0c0;
+    color: #C1C1C1;
+    font-feature-settings: 'liga' off, 'clig' off;
+    font-family: Pretendard;
+    font-size: 17px;
+    font-style: normal;
+    font-weight: 500;
+    line-height: 140%; /* 23.8px */
+    letter-spacing: -0.425px;
   }
 
   > input:focus {
