--- src/pages/SearchPage/SearchPlace.tsx
@@ -1,29 +1,0 @@
-import React, { useState } from "react";
-
-import Header from "../../components/forSearchPage/Header/Header";
-import Navigation from "../../components/Navigation/Navigation";
-
-import KakaoMap from "../../components/forSearchPage/KakaoMap/KakaoMap";
-import Places from "../../components/forSearchPage/Places/Places";
-
-export default function SearchPage() {
-  const [currentLocationName, setCurrentLocationName] = useState<string>("");
-  const [searchAddress, setSearchAddress] = useState<string>("");
-  
-  return (
-    <>
-      <Header 
-        placeholder={currentLocationName} 
-        onSearch={setSearchAddress}
-      />
-      <KakaoMap 
-        mapHeight="350px" 
-        setCurrentLocationName={setCurrentLocationName}
-        searchKeyword={searchAddress}
-        buttonOn={true}
-      />
-      <Places />
-      <Navigation />
-    </>
-  );
-}
